Chat-Client-Server-Program-using-GUI-JAVA-
Write  a  program  that  enables  two users  to chat.  Implement  one user  as  the  server   and  the other as the client. The server has two text areas: one for entering text and the other (noneditable) for displaying text received from the client. When the user presses the Enter key, the current line is sent to the client. The client has two text areas: one for receiving text from the server, and the other for entering text. When the user presses the Enter key, the current line is sent to the server. Name the client Exercise25_10Client and the server Exercise25_10Server.You may write the program using datagram sockets